+++ src/test/java/dev/orne/test/rnd/params/GenerationParametersExtractorTest.java
@@ -61,6 +61,9 @@ extends BaseParametersSourceExtractorTest {
     @Test
     void testSupportedParametersTypes() {
         assertSupportsParametersType(GenerationParameters.class);
+        assertSupportsParametersType(StringGenerationParameters.class);
+        assertSupportsParametersType(CollectionGenerationParameters.class);
+        assertSupportsParametersType(MapGenerationParameters.class);
     }
 
     /**