+++ src/test/java/dev/orne/test/rnd/params/NotNullConstraintExtractorTest.java
@@ -58,6 +58,7 @@ extends BaseParametersSourceExtractorTest {
         assertSupportsParametersType(GenerationParameters.class);
         assertSupportsParametersType(StringGenerationParameters.class);
         assertSupportsParametersType(CollectionGenerationParameters.class);
+        assertSupportsParametersType(MapGenerationParameters.class);
     }
 
     /**