+++ src/test/java/dev/orne/test/rnd/params/SizeConstraintExtractorTest.java
@@ -65,6 +65,7 @@ extends BaseParametersSourceExtractorTest {
         assertSupportsParametersType(SizeParameters.class);
         assertSupportsParametersType(StringGenerationParameters.class);
         assertSupportsParametersType(CollectionGenerationParameters.class);
+        assertSupportsParametersType(MapGenerationParameters.class);
     }
 
     /**